Split linkVariants into its own file
