+++ packages/html-reporter/src/links.tsx
@@ -14,15 +14,7 @@
   limitations under the License.
 */
 
-import type { TestAttachment, TestCase, TestCaseSummary, TestResult, TestResultSummary } from './types';
 import * as React from 'react';
-import * as icons from './icons';
-import { TreeItem } from './treeItem';
-import { CopyToClipboard } from './copyToClipboard';
-import './links.css';
-import { linkifyText } from '@web/renderUtils';
-import { clsx, useFlash } from '@web/uiUtils';
-import { trace } from './icons';
 
 export function navigate(href: string | URL) {
   window.history.pushState({}, '', href);
@@ -38,7 +30,7 @@ export const Route: React.FunctionComponent<{
   return predicate(searchParams) ? children : null;
 };
 
-type LinkProps = React.PropsWithChildren<{
+export type LinkProps = React.PropsWithChildren<{
   href?: string,
   click?: string,
   ctrlClick?: string,
@@ -55,75 +47,6 @@ export const Link: React.FunctionComponent<LinkProps> = ({ click, ctrlClick, chi
   }}>{children}</a>;
 };
 
-export const LinkBadge: React.FunctionComponent<LinkProps & { dim?: boolean }> = ({ className, ...props }) => <Link {...props} className={clsx('link-badge', props.dim && 'link-badge-dim', className)} />;
-
-export const ProjectLink: React.FunctionComponent<{
-  projectNames: string[],
-  projectName: string,
-}> = ({ projectNames, projectName }) => {
-  const encoded = encodeURIComponent(projectName);
-  const value = projectName === encoded ? projectName : `"${encoded.replace(/%22/g, '%5C%22')}"`;
-  return <Link href={`#?q=p:${value}`}>
-    <span className={clsx('label', `label-color-${projectNames.indexOf(projectName) % 6}`)} style={{ margin: '6px 0 0 6px' }}>
-      {projectName}
-    </span>
-  </Link>;
-};
-
-export const AttachmentLink: React.FunctionComponent<{
-  attachment: TestAttachment,
-  result: TestResult,
-  href?: string,
-  linkName?: string,
-  openInNewTab?: boolean,
-}> = ({ attachment, result, href, linkName, openInNewTab }) => {
-  const [flash, triggerFlash] = useFlash();
-  useAnchor('attachment-' + result.attachments.indexOf(attachment), triggerFlash);
-  return <TreeItem title={<span>
-    {attachment.contentType === kMissingContentType ? icons.warning() : icons.attachment()}
-    {attachment.path && (
-      openInNewTab
-        ? <a href={href || attachment.path} target='_blank' rel='noreferrer'>{linkName || attachment.name}</a>
-        : <a href={href || attachment.path} download={downloadFileNameForAttachment(attachment)}>{linkName || attachment.name}</a>
-    )}
-    {!attachment.path && (
-      openInNewTab
-        ? (
-          <a
-            href={URL.createObjectURL(new Blob([attachment.body!], { type: attachment.contentType }))}
-            target='_blank' rel='noreferrer'
-            onClick={e => e.stopPropagation() /* dont expand the tree item */}
-          >
-            {attachment.name}
-          </a>
-        )
-        : <span>{linkifyText(attachment.name)}</span>
-    )}
-  </span>} loadChildren={attachment.body ? () => {
-    return [<div key={1} className='attachment-body'><CopyToClipboard value={attachment.body!}/>{linkifyText(attachment.body!)}</div>];
-  } : undefined} depth={0} style={{ lineHeight: '32px' }} flash={flash}></TreeItem>;
-};
-
-export const TraceLink: React.FC<{ test: TestCaseSummary, trailingSeparator?: boolean, dim?: boolean }> = ({ test, trailingSeparator, dim }) => {
-  const firstTraces = test.results.map(result => result.attachments.filter(attachment => attachment.name === 'trace')).filter(traces => traces.length > 0)[0];
-  if (!firstTraces)
-    return undefined;
-
-  return (
-    <>
-      <LinkBadge
-        href={generateTraceUrl(firstTraces)}
-        title='View Trace'
-        className='button trace-link'
-        dim={dim}>
-        {trace()}
-        <span>View Trace</span>
-      </LinkBadge>
-      {trailingSeparator && <div className='trace-link-separator'>|</div>}
-    </>
-  );
-};
-
 export const SearchParamsContext = React.createContext<URLSearchParams>(new URLSearchParams(window.location.hash.slice(1)));
 
 export const SearchParamsProvider: React.FunctionComponent<React.PropsWithChildren> = ({ children }) => {
@@ -137,64 +60,3 @@ export const SearchParamsProvider: React.FunctionComponent<React.PropsWithChildr
 
   return <SearchParamsContext.Provider value={searchParams}>{children}</SearchParamsContext.Provider>;
 };
-
-function downloadFileNameForAttachment(attachment: TestAttachment): string {
-  if (attachment.name.includes('.') || !attachment.path)
-    return attachment.name;
-  const firstDotIndex = attachment.path.indexOf('.');
-  if (firstDotIndex === -1)
-    return attachment.name;
-  return attachment.name + attachment.path.slice(firstDotIndex, attachment.path.length);
-}
-
-export function generateTraceUrl(traces: TestAttachment[]) {
-  return `trace/index.html?${traces.map((a, i) => `trace=${new URL(a.path!, window.location.href)}`).join('&')}`;
-}
-
-const kMissingContentType = 'x-playwright/missing';
-
-export type AnchorID = string | string[] | ((id: string) => boolean) | undefined;
-
-export function useAnchor(id: AnchorID, onReveal: React.EffectCallback) {
-  const searchParams = React.useContext(SearchParamsContext);
-  const isAnchored = useIsAnchored(id);
-  React.useEffect(() => {
-    if (isAnchored)
-      return onReveal();
-  }, [isAnchored, onReveal, searchParams]);
-}
-
-export function useIsAnchored(id: AnchorID) {
-  const searchParams = React.useContext(SearchParamsContext);
-  const anchor = searchParams.get('anchor');
-  if (anchor === null)
-    return false;
-  if (typeof id === 'undefined')
-    return false;
-  if (typeof id === 'string')
-    return id === anchor;
-  if (Array.isArray(id))
-    return id.includes(anchor);
-  return id(anchor);
-}
-
-export function Anchor({ id, children }: React.PropsWithChildren<{ id: AnchorID }>) {
-  const ref = React.useRef<HTMLDivElement>(null);
-  const onAnchorReveal = React.useCallback(() => {
-    ref.current?.scrollIntoView({ block: 'start', inline: 'start' });
-  }, []);
-  useAnchor(id, onAnchorReveal);
-
-  return <div ref={ref}>{children}</div>;
-}
-
-export function testResultHref({ test, result, anchor }: { test?: TestCase | TestCaseSummary, result?: TestResult | TestResultSummary, anchor?: string }) {
-  const params = new URLSearchParams();
-  if (test)
-    params.set('testId', test.testId);
-  if (test && result)
-    params.set('run', '' + test.results.indexOf(result as any));
-  if (anchor)
-    params.set('anchor', anchor);
-  return `#?` + params;
-}
